+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - added `depositLimit()`
 - renamed `rate()` to `exchangeRate()`
 - added `recent()`
+- intercept http methods
 
 0.0.1 / 2015-04-13
 ------------------